requirements.txt
Babel==2.12.1
blinker==1.5
certifi==2022.12.7
click==8.1.3
dnspython==2.3.0
elastic-transport==8.4.0
elasticsearch==8.6.2
email-validator==1.3.1
Flask==2.2.3
flask-babel==3.0.1
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==5.1.1
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mongoengine==0.27.0
passlib==1.7.4
pymongo==4.3.3
python-dotenv==1.0.0
pytz==2022.7.1
urllib3==1.26.14
Werkzeug==2.2.3
WTForms==3.0.1
